panel/02-contributors_graph.py: Notify about demo data using notifications
Instead of using the Alert pane to notify about the fact that the app
did not find data files it needs, use the notification mechanism.

Note that the app must be set up, and fully loaded, to be able
to display notification.  That is why the info notification is
sent from the onload callback.

The notification in question, as currently configured, stays forever.
The user can close it using the "X" button, however.

.
While at it, add a few print statemet for debugging (commented out).