panel/02-contributors_graph.py: Use <time> element for dates
This required switching from pn.pane.Markdown to pn.pane.HTML, but
the look did not change.

Also, try harder to be able to use day of the month as a decimal number
rather than as zero-padded decomal number (e.g. 8 instead of 08); in
Python it is unfortunately platform specific (%-d or %#d vs %d).