panel/02-contributors_graph.py: Add warning_notification()
This is very simple ad-hoc solution to the problem that the Panel
notification system (`pn.state.notifications`) only starts working after
application is fully loaded.  If you try to display warning or error
before then, it never appears.

The warning_notification() class checks the `loaded` variable, and if
application is not yet loaded, saves the message one the "stack".  If
application is loaded, it uses pn.state.notification.

Warnings on the stack are picked and displayed by the onload callback.

.
An alternative would be to create proxy object for the
pn.state.notification, and save delayed functions (see 'joblib')
to the object state.