panel/02-contributors_graph.py: Beginnings of authors grid
Currently top N authors (based on selected contributions criteria, be it
commits, additions, or deletions) are selected to be displayed, and just
author e-mails are displayed in 2-column grid.

The N in top N can be selected with a widget (default is to show 4
authors).

The grid is reactive: it changes when changing the dataset / repo, when
changing N in "top N", and when changing the type of contribution.
Currently it DOES NOT take into account "Period:" (i.e. date range).

NOTE that trying pn.rx() did not work, and pn.bind() to be used.