panel/02-contributors_graph.py: Use per-author plots (separate scales)
Each of per-author Cards in the 2-column authors grid now show
appropriate per-author plot.  Currently all those plots, together with
the plot for the whole repository, at first have the same axes ranges,
but they may decouple after changing the state of controlling widgets,
for example the "top N" widget - this is TO BE FIXED in one of the
subsequent commits.

.
The error that I stumbled upon in previous commit was not cause, as I
thought, by the `column` parameter not being of the 'str' type, but the
fact that for some strange reason doing groupby() then resample(), like
in the following code:

    df_x.groupby('author.email').resample(
        resample_rate,
        on='author_date'
    )[columns_agg].agg(
        agg_func_map,
        numeric_only=True
    )

results in a dataframe that has hierarchical *columns* (cross-product of
the `columns_agg` columns), thus names of columns are not 'str' but
'tuple'.  Each of those groups in hierarchical columns are repeats.

                      n_commits	|                       +:count	| ...
  n_commits | +:count | -:count	| n_commits | +:count | -:count | ...

However, doing groupby() and resample() in a single operation, like in
the following code:

    df_gr = df_x.groupby([
        'author.email',
	pd.Grouper(key='author_date', freq=resample_rate)
    ]).agg(
        agg_func_map,
        numeric_only=True
    )

results in a DataFrame wit only 'n_commits', '+:count', and '-:count'
columns, without this hierarchical index for columns stuff.