plot/02-contributors_graph.py: Change plot color depending on `column`
Use blue for "Commits", green for "Additions", and red for "Deletions".
Trying to use colors with the same lighness / brightness via OKLCH color
schema.

Alternative would be to use the same colors as in per-author info for
number of additions and deletions, namely #1a7f37 (rather than #008826)
for additions, and #d1242f (rather than #d42000) for deletions.