panel/02-contributors_graph.py: Histogram on margin of main plot
Add toggle-able histogram of values on the right marging of the main
plot (for the whole repository).

BUG: Currently when toggling marginals on, the main plot no longer
stretches full width available to it, but takes only around half of the
available width.

TODO: configure number of bins, maybe make bins fall on "natural"
numbers, like 2, 5, 10, etc., like nice() or rangeRound() in D3.js.