notebooks/panel: Add throttling to 00-panel-simple-Matplotlib.ipynb
Use `xrange_slider.param.value_throttled` instead of `xrange_slider`
(and similarly for `yrange_slider` in definition of `bound_plot_streamplot`.
This follow the first suggestion from "Enable throttling" subsection in How-to:
https://panel.holoviz.org/how_to/performance/throttling.html

However, without the next part of the change, the "Reset" button stops
working fully: it still resets sliders, but do not reset plotted range.
One solution is to use the following code isnide button callback:

    with param.edit_constant(xrange_slider):
        xrange_slider.value_throttled = xrange_slider.value = (-3, +3)

https://discourse.holoviz.org/t/throttling-pn-bind-and-using-button-to-reset-values/8341/5

There is no description of `param.edit_constant` in Panel documentation
(thougt it is used in example code in "Set Up manual Throttling" and in
"Streaming Videostream").  It is described only in Param documentation:

https://param.holoviz.org/user_guide/Parameters.html
https://param.holoviz.org/reference/param/generated/param.edit_constant.html