02-contributors_graph.ipynb: Format hvPlot part of "-/+ plots" - spacing
When trying to create equivalent of the following line with Matplotlib:

   fig.subplots_adjust(hspace=0)

it turned out that the solution depends on the backend used.

The `vspace=0` or `tight=True` options are available for hv.Layout
only for the 'matplotlib' backend, but support for them is missing
from the 'bokeh' backend (this feature is missing at least since 2020:
https://discourse.holoviz.org/t/is-vspace-and-hspace-specific-to-matplotlib-backend-is-there-a-bokeh-equivalent-how-to-stack-plots/502)

As a workaround, one can either use `border=0` option to *plot*
available for 'bokeh' backend, compose result using bokeh.layouts
(e.g. bokeh.layouts.gridplot) instead of using hv.Layout, etc.
https://www.perplexity.ai/search/translate-the-following-code-i-Dz4RBrHFTl.FrUP_FmP84w

The solution in this commit uses `border=1` / `border=0` "trick".