02-contributors_graph.ipynb: Draw area for "-/+ plots", begin "Sankey"
In the "hvPlot" subsection of the "-/+ plots" section creating '-:count'
plot and '+:count' plot is now separated from the process of putting them
together on a single figure.

This allowed creating a workaround for HoloViews lacking support for
`interpolate` option to hv.Area like there is one for hv.Curve
https://holoviews.org/reference/elements/bokeh/Curve.html#interpolation
and hvplot.area lacking support for `step` option, like there is one
for ax.fill_between in Matplotlib.  This support is a bit of a hack
in that it duplicates the amount of data points, creating a data for
a step function - currently only when='post' is supported - and then
running hvplot.area on this augmented data.

But it seems to work all right, both for normal and for inverted plot,
though surrectly there is no support for filling area between two curves:
only filling area batween curve and the Ox axis is implemented.

.
Unfortunately intermingled with the rest of changes are beginnings
of the "Sankey" section (I forgot to commit changes before starting
a new topic).