02-contributors_graph.ipynb: Add "Larger single commit example - automatic" subsection
This subsection takes the code for creating data and refining Sankey
plots from the "Single commit example" subsection, creates functions
from them, and applies to changes from a commit in qtile/qtile with
the largest number of changed files.

This produces the same evolution of Sankey plots, though because
of the large amount of changed files the first two plots (with files
and file names) are night unreadable.