02-contributors_graph.ipynb: Add "Styling Sankey (single large commit, processed)"
This section switches to using DataFrame as the data source, and starts
to configure how the Sankey diagram would look like.

Adds a few references to take a look into, on how to style and shape Sankey plots.

NOTE: With Bokeh (and possibly also with Plotly) it is not possible to
have color gradient along the edge in the graph, unfortunately (or not).