diffinsights_web/02-contributors_graph.py: Tab with JSON and Perspectives
Help with finding problems by providing, at the bottom of the page,
the Tabs layout, allowing to switching between showing raw input data
in a JSON format, input data as DataFrame (using Perspective pane),
resampled data, and grouped by author and resample data.

For example, one can see that the problem with per-author plots skipping
data points is caused by groupby time values missing (or at least that is
what Perspective pane shows for groupby result), not because of N/A
values.  So it is maybe a question of adding `dropna=True` to groupby(),
rather than adding .fillna(0) after groupby() and agg().

Perhaps something like that (maybe with Tabulator instead, or in
addition to Perspective) would be a better replacement for "Save as CSV"
capability of GitHub Insights.