02-contributors_graph.ipynb: Add "Un-sparseify time series" section
This section deals with solving the issue noted in diffinsights_web/TODO.md:

> In per-author plot, missing values creates "jumps" instead of being
> filled with 0;

This is caused by the fact that double groupby() sparsifies time series,
omitting entries if there is nothing there.  That is not good whe using
this data for plotting.

In short: .resample(<original resample rate>).first(), or simply
.asfreq(<original resample rate>), possibly followed by .fillna(0)
will solve this issue.