diffinsigths_web: Add apps/contributors.py
Start splitting what was originally notebooks/panel/02-contributors_graph.py,
and is now diffinsights_web/02-contributors_graph.py, making it more
modular, and more readable and easier to maintain.  This split will
follow the ideas in "Structure with a DataStore" intermediate level
tutorial from the Panel documentation:
  https://panel.holoviz.org/tutorials/intermediate/structure_data_store.html

While at it, add "web" optional dependency to pyproject.toml, currently
just with 'panel' in it.

Note that the new app (that currently does not actually do anything),
can be run with `panel serve src/diffinsights_web/apps/contributors.py',
but also with `python src/diffinsights_web/apps/contributors.py` (as
a script), though it is not preferred way of doing it.  Note that
those different ways of running the app result in different local URLs
it is at.

TODO: add a README.md for the [new] app.