diffinsights_web: Add (and check) utils/notifications.py
The code to display notification regardless of whether it was send
before there was place to display it, or after it, was taken from
02-contributors_graph.py.  Some minor changes had to be made for
it to work as intended (i.e. the warning should be shown on each
reload).

The check was manual, and not automatic, done by running the app
and seeing if the warning will appear.

.
NOTE that `panel serve` runs with files from source, not from installed
package; that is why diffinsights_web/apps/contributors.py requires
absolute imports, while we could use relative imports in diffannotator/*.py
scripts.

.
By the way, it turns out that there is regression in Panel 1.5.4
in regards to notifications, detected when testing this change:
https://github.com/holoviz/panel/issues/7504