diffinsights_web: Add button for clearing the cache
This could be helpful while developing the app, to avoid situation where
the app shows old look because of the caching.

New widget was added to widgets/, instead of putting it in views/;
because the purpose is different, but it could have gone other way.

We should probably inform about cache getting cleared, but there
is only support for warnings, and no support for info messages in
utils/notifications.py