diffinsights_web: Create DataFrame from JSON data, and display it
Extract data for selected repository from the JSON data, and create
pd.DataFrame from it.  Display this data for ad-hoc analysis, copying
and saving, by using the pn.pane.Perspective.

Currently the new view / pane is added simply at the bottom
of the page, below the JSON viewer.

While at it, store all widgets used by the TimelineDataStore class
in instance variables, and not only on self._widgets list (which is
used by the __panel__() method).

.
This change required adding 'pandas' to the [web] requirements
in pyproject.toml.