diffinsights_web: Attempt at creating per-class widget
Use param.ClassSelector with instantiate=False and per_instance=False
inn a failed attempt to have single instance of frequency selector
for every instance of ResampledTimelineDataStoreClass (every object
of this class).

This attempt failed: there are still two selector present, each with
its own value.  It is possible that it can be done with 'param' library,
but next commit will take a simpler approach.

For comparison, temporary add the same selector twice to sidebar.