diffinsights_web: Add "±:<sth> [%]" columns to resampled df (WIP)
Find all columns that start with "+:" or "-:", like "+:type.code",
and add [%] version of that column to the dataframe after resampling.

This makes sense only if we are using the 'sum' aggregation function,
like it is done now (but there are plans to make it configurable).

This change is unfinished.  There are a few things that needs to be
corrected:

1. Adding [%] columns must be done in one step with creating both
   dataframes with resampled (and aggregated) data, so that it is
   correctly recalculated if input dataframe changes (e.g. if user
   selects different input JSON file).

   This will also fix the problem with the current code being buggy:

      AttributeError: The value of a derived expression cannot be set.
      Ensure you set the value on the root node wrapping a concrete
      value, e.g.:

          a = rx(1)
          b = a + 1
          a.rx.value = 2

      is valid but you may not set `b.rx.value = 2`

2. If the value in "+:count" column is 0, then any "+:<column> [%]"
   value is undefined.  Mark it so explicitly with np.nan
   (or equivalent).

3. For every "-:<column>" and "+:<column>", create "±:<column>"
   (or just "<column>") column that is sum of "-:<column>" and
   "+:<column>".

   Note that ("-:count" + "+:count") can be 0 if there are only
   changes to binary files (with no lines change), though such
   case should be rare.

After funishing this change we should be able to see new columns
in the appropriate Perspective panes (at the bottom of the web page).

Next step: enable plotting [%]-derived plots.  And remove debug-print
statements...