diffinsights_web: Move add_pm_count_perc() call into resample_timeline()
This way the [%] columns will be recomputed at the same time when
resampling and aggregation is done.

This gets rid of the error that was shown in previous commit.  Now
[%] columns are visible in relevant Perspective panes.  Also, when
switching to different JSON file (and to different repo), [%] columns
are there after change.

Next will be removal of debug-print statements, and adding cumulative
baror area plot for [%] columns.