diffinsights_web: Add "Line types distribution +:[%]" plot
Currently this plot is a cumulative / stacked area plot, because
it looks like cumulative / stacked bar plot does not work - whole
plot is just one black area for some reason.

Label is put on the bottom, though it has default description:
"Variables" for column names.  Columns were not sorted, so the
order is a bit random; colors are also not the best (perhaps
we should use colormap).

The area plot in hvPlot has the same limitations as stackplot()
in matplotlib; it does not allow the use of step interpolation.
Side note: on HoloViews hv.Curve has `interpolation` option that
can take 'linear', 'steps-mid', 'steps-pre', and 'steps-post'
values - but there is no such option for hv.Area (I don't know why?).
There is nothing like `drawstyle='steps-post'` for sns.lineplot
and `step='post'` for ax.fill_between (in seaborn and Matplotlib,
respectively).

The special case code is a bit fragile: there are several places
where the name (currently: "KIND [%]") must be given.  This needs
to be improved.