diffinsights_web: Add RepoPlotHeader above TimelinePlot
On the main part of the apps/contributors.py page, add a header with
the information about type of contributions, resampling frequency,
and displayed date range (this takes into account the value of "Period:"
widget by checking the displayed range of the plot).

This is almost straight copy of the code from 02-contributors_graph.py.