diffinsights_web: Add 'selected author' pane to the tabbed widget
Add the DataFrame with resampled information for a single selected
author to the tabbed widget at the bottom ot the main part of the page,
via pn.pane.Perspective pane.

The code is currently present in apps/contributors.py, but it will be
(at least in part) extracted to a separate file.

Manual exploration shows that there are entries missing (skipped
periods, for example skipped months for 'ME' / 'month end' resample
frequency) in this per-author DataFrame.

This code was copied, with changes, from 02-contributors_graph.py.