diffinsights_web: Temporarily remove authors_info_panel ('selected author')
It turned out that the code extracted out of 02-contributors_graph.py
doesn't work with changing the input JSON file, failing with the
following error:

  panel.reactive - Callback failed for object named 'input JSON file'
                   changing property {'value': ...}
  [...]
  KeyError: 'mcol@posteo.net'

The traceback includes the following reference to the app code:

  Traceback (most recent call last):
  [...]
  File "python-diff-annotator/src/diffinsights_web/datastore/timeline.py",
  line 207, in author_timeline_df
    return resample_by_author_df.loc[author_id]

One can check that after changing the value of the "input JSON file"
widget, the "selected author" tab in tabbed widget has its selector
widget with the author from the JSON file before, and old dataframe.

As this is not a critical component (though it would be useful for
debugging), instead of taking time to tyring to find the source of this
error - even though the code looks correct - comment out and remove
this component / widget.

TODO: find the source of this error, and fix the bug.