diffinsights_web: Add per-author plots for each author in the grid
Uses TimeseriesPlotForAuthor class to add per-author plot (same as
the main plot for the repo) for each author in the top N authors grid.

NOTE that the problem with the fact that per-author dataframe can have
some of its periods missing, instead of being filled with 0s or NaNs.
To be fixed in some future commit, see "Un-sparseify time series" section
in notebooks/panel/02-contributors_graph.ipyb