diffinsights_web: Fill gaps in frequency for per-author plots
The resample+groupby process creates sparsified dataframe, with periods
in which there are no samples entirely missing, instead of having one
entry per given frequency, even if some of them would be all zeros.

This however matters for plots, so fill the dataframe with per-author
data with zeros, if needed, before plotting.  The author_timeline_df_freq()
function was created for this purpose.  It is based on experiments in
the "Un-sparseify time series" section:
  notebooks/panel/02-contributors_graph.ipynb#Un-sparseify-time-series

TimeseriesPlotForAuthor now uses this function (with the help of pn.rx()).

While at it, comment-out remaining debug-print command.