diffinsights_web: Add ContributionsPercHeader for per-repo plot
This adds indicator for split between different line types in changed
lines (both deleted and added lines).  This indicator is similar to
the "Languages" information in the sidebar on the right side of the
repo view on GitHub.

Currently it does not yet include the legend for the indicator,
beside the use of the "title" attribute.

The look of the indicator was developed using CodePen:
https://codepen.io/jnareb/pen/jOgQYxg

.
TODO: fix a bug with the indicator not refreshing after changing
the "Period:" widget value, as it should (similar code does work
correctly in other places).