diffinsights_web: Add workaround for author_timeline_df{,_freq} bug
Add a workaround for the issue first noticed and notified in commit
6abc649 (diffinsights_web: Temporarily remove authors_info_panel
('selected author'), 2024-11-25) for author_timeline_df() function,
but actually present for author_timeline_df_freq() until now.

The problem seems to be caused by some stale unreaped object or reactive
function invoking author_timeline_df() or author_timeline_df_freq() with
the new value for resample_by_author_df - because it is reactive
expression, and old value from previous DataFrame for author_id - because
it is static.

The **WORKAROUND** is to check whether `author_id` is present in the
`resample_by_author_df` DataFrame index, and if it is not, warn and use
simply first author in the provided `resample_by_author_df` DataFrame.

Now, after switching from one JSON data file to another, we get the
following series of messages:

  WARNING: author_timeline_df_freq(): author_id='mcol@posteo.net' not in
  resample_by_author_df=<0x186feb6ef50> index
           using author_id='ezhulenev@google.com' instead
  [...]
  2024-11-26 17:30:30,018 Dropping a patch because it contains a previously
  known reference (id='p1741'). Most of the time this is harmless and usually
  a result of updating a model on one side of a communications channel while
  it was being removed on the other end.

.
TODO?: Undo removal of the Perspective pane tab with DataFrame with data
for selected author.  It should now work thanks to this workaround.