annotate.py: Add 'line_file_no' field to annotation data
This field tracks the line number in the original file that the line in
the changeset came from, namely the line number in the source file in
case of removed lines ('-'), and the line number in the target file in
case of added lines ('+') and context lines (' ') - the last one is not
used by the annotation process.

This change would make it easier to compare anotation from the
python-diff-annotate tool (from the diffannotator package), with
automatic or manual annotation in other line annotation datasets,
like our own HaPy-Bug, or SmartSHARK dataset (Herbold et al. 2022).