01-compare_annotations.ipynb: Add "Running annotation on BugsInPy repos", etc.
Find where repositories are, and create strings with command line invocations
of `diff-annotate from-repo ...` with output dir, and with repo and commits.

When starting to compare collective.csv data with created line annotations
it turned out that `diff-annotate` was run with default settings, which
*did not* use file purpose for setting line type with --purpose-to-annotation
options.

TODO: redo the annotations (to save work, maybe create script file to run).