data/experiments: Add HaPy-Bug/{bip,cve,crawl}_blame.csv
Add bip_blame.csv, cve_blame.csv and crawl_blame.csv to the HaPy-Bug/
subdirectory in data/experiments/.

Those files have the following structure (the following columns):

  ,bug_id,commit,repo,filename,lineNumber,start time,end time

in the case of bip_blame.csv, and

  ,bug,commit,repo,filename,lineNumber,blame,rev blame

in the case of cve_blame.csv and crawl_blame.csv.

The data in those files can be used to find appropriate commit for the
given bug (in BugsInPy, aka bugs-in-py, aka bip), or CVE (in cve and
crawl).

The files were send by Aleksander Kazecki, one of co-authors of the
"HaPy-Bug - Human Annotated Python Bug Resolution Dataset" draft paper.