notebooks/experiments: Add --purpose-to-annotation to command args
Adjust instructions to generate annotations for the BugsInPy subset
of the HaPy-Bug dataset to include the following global options:

    --purpose-to-annotation=data \
    --purpose-to-annotation=documentation \
    --purpose-to-annotation=markup \
    --purpose-to-annotation=other \
    --purpose-to-annotation=project \
    --purpose-to-annotation=test \

This should come closer to how the old version of the automatic annotation
process used as pre-processing step in creation of HaPy-Bug dataset.

Adds run_annotation_bugsinpy_repos.sh script created by the notebook,
adding it in the data/experiments/HaPy-Bug/ directory.