annotate.py: Use "code" only for files with "programming" purpose
The default annotation rule was to mark line with file purpose if that
purpose was present in PURPOSE_TO_ANNOTATION, or "documentation" if
it passed line_is_comment() test, otherwise "code".

The line_callback could be used to override the default rule.

This commit changes this default rule, so that it returns "code" only
if teh file_purpose was "programming", and it returns file_purpose
otherwise.

Intruduces purpose_to_default_annotation().