annotate.py: Fix line_is_comment() logic, and fix its test
The line_is_comment() logic was to be that line that consists only
of comments and whitespace should be considered comment only.  For
example, indented single-line comment should be considered comment-only
line, even though it has some whitespace tokens.

However, the code of line_is_comment() returned True also on line
which was whitespace-only (which includes empty lines).  This should
not be the case.

Fix the code, and improve the test so that it would catch the bug.