annotate.py: Add `file_purpose` parameter to line callback
The LineCallback type now includes second parameter, and line callbacks
are called with file purpose (str) as its first parameter, and tokens
list only as second.

Note that when providing just the body of the line callback, you *must*
use the `file_purpose` and `tokens` as [parameter] names; when providing
the code of the function, then first positional parameter must be file
purpose, and second positional parameter must be the list of tokens.

TODO: document line callback functions better; see e.g. git-filter-repo.

NOTE that this are of code is not well tested, so there can be some
bugs that were not detected.