01-compare_annotations.ipynb: Add comparison between manual and auto
Adds "Extract post-processed auto" and "Compare manual with auto (from
collective)" subsections, that try to compare the consensus between human
annotators, and automatic annotations before human changes.

This comparison is limited to the BugsInPy ($D_{BIP}$) subset of the
whole HaPyBug dataset.

The code is copied from and/or based on the code in 00-HaPy_Bug-Paper.ipynb,
namely the "Consensus" section and its "Consensus vs Auto" subsection.

NOTE that this notebook was *not* re-run after changes to annotate.py,
and after re-running run_annotation_bugsinpy_repos.sh script (which runs
with output data outside the Git repo).