annotate.py: Change `file_purpose` to `file_data` in line callback
The LineCallback type now uses the `dict[str, str]` and not `str` as the
type of its first parameter; the line callback is passed file_data, not
just file_purpose as its first argument (positional argument).

This change allows to modify code of hapybug_line_callback_func.py
to match what the code to create initial automatic annotations for
the annotation tasks in the HaPy-Bug dataset does.

Renames `file_purpose` to `file_data`, and use `file_data["purpose"]`
in place where there was `file_purpose` previously.

Test passes, manual checking with the help of test and debug-print
statements suggests that it should work as intended.