Revert "diffinsights_web: Temporarily remove authors_info_panel ..."
This reverts commit 6abc6496f8fe00302d76b3e29c1f9f9812e250a9.

While at it, optimize imports in contributors.py, and add 2 to the list
of possible values for "top N" widget in authorsgrid.py (so that this
option doesn't vanish due to this revert).

The authors data pane that was brought back doesn't work entirely
relaibly; after changing JSON file, the widget for selecting the
authors shows old list of authors.