01-compare_annotations.ipynb: Compute hapy_bit_repos_vs_consensus, do spot-check
The hapy_bit_repos_vs_consensus will be used to find out what is the
cause of difference between `diff-annotate dataset` and `diff-annotate from-repo`
results with regard to disagreement with the consensus of human annotations
in HaPy-Bug dataset (BugsInPy subset).

Do a bit of spot check, examining the cases where there is difference
between consensus, and the result of running `diff-annotate from-repo`
with the help of run_annotation_hapy_bip_repos.sh script (run by hand,
not by the notebook).