Add various data/examples/stats/hellogitworld.*.json files
Those files are the result of annotating (with `diff-annotate from-repo`)
and analyzing (with `diff-gather-stats ...`) commits in a small example
repository: https://github.com/githubtraining/hellogitworld.git

They were added to have something to work with for a demo of web app
(of a dashboard, visualizing different metrics and statistics), even
if nobody ran `dvc repro` or `dvc pull`.

The hellogitworld repository was a sample project students can use during
Matthew J. McCullough's Git class.  It has been archived by the owner on
Jan 24, 2020, and it is now read-only.