diffinsights_web: Switch from DD.MM.YYYY to YYYY.MM.DD for from_date_str
GitHub Insights use the DD.MM.YYYY format for the 'from' parameter
storing the value of the "Period:" widget, see for example the URL
https://github.com/qtile/qtile/graphs/contributors?from=26.11.2022

That's why this was the format used in PatchScope web app.  However,
the YYYY.MM.DD has the advantage that lexical sorting (sorting it as
a string) gives the same result as chronological sorting (sorting it
like a date).

Note that this change uncovered a few places with code duplication.