01-compare_annotations.ipynb: Run comparison between SmartSHARK and diff-annotate
Improve a bit script to run `diff-annotate from-repo` on SmartSHARK repositories.
This script is saved as run_annotation_smartshark_repos.sh in the
data/experiments/SmartSHARK/ directory... but it was not run from there.
Line callback prepared for creating SmartSHARK-like annotations is
available in the same directory as smartshark_line_callback.py.

Run the script (after changes to annotate.py in previous commits), and
retrieve the annotation data, and put it into DataFrame.

Two DataFrames: from SmartSHARK and from diff-annotate are compared with
each other with the help of pd.merge.  The results were (lightly) analyzed.

NOTE: there was no spot-check to examine the discrepancies, and why there
were so many mis-matches.  Maybe the pairig process is actually faulty.