01-compare_annotations.ipynb: Examine 'hunk_id', single-hunk limited comparison
In face of large mismatch between SmartSHARK annotations, and `diff-annotate`
annotations generated on SmartSHARK repos and commit, the notebook includes
examination of what 'hunk_id' might mean.

It turns out that it is neither related to hunk number in a patched file
in a diff (there are more different hunk_id's for a sha+file than that),
nor to the change group i.e. group of consecutive changed lines in a patched
file (for diff with a single changed file, there are more change groups
than different hunk_id's).

This means that we cannot identify matching lines because of the broken
link in the chain <project> -> <revision> -> <file> -> ... -> <line_idx>.

Therefore, let's examine the "single hunk" case, i.e. cases where there
should be only a single 'hunk_id' for changed file in a revision, and
patched files with only a single hunk (on the `diff-annotate` side).

Still around 30% of mismatch (if much less lines withou match - they
did not vanish, but there is less of them), so let's adjust annotation
rules.