diffinsights_web: Count "Period:" from latest date in dataframe
Instead of counting the period, like "Last 3 months", from todays date,
count it from the latest 'author_date' in the dataframe.

This required adding a new param to ContributorsHeader class, and
adding new reactive field to the TimelineDataStore class.

Only lightly tested (manually).

Note that this does not get rid of all problems with respect to empty
result of filtering by date, so the guard is still needed.