02-compare_annotations_Herbold.ipynb: More about missing repos
In Herbold et al. dataset, for 2 projects out of 29, we were not able
to run `diff-annotate from-repo ...` on those repositories.  It turns
out this is caused by the fact that revision assigned in Herbold et al.
to those repositories are actually missing from those repositories,
at least that is the case for a few SHAs that were checked.

Count how many commits those 2 projects / repositories have in Herbold et al.
dataset, and how many annotated lines.