01-compare_annotations.ipynb: Start examining causes of lack of match
Beginnings of investigation why there are lines for which there is no match
between lines in the original dataset and lines from `diff-annotate ...`.
For now, let's begin with examining the case for BugsInPy from HaPy-Bug.

It turns out that one of the problems is that `diff-annotate from-repo ...`
drops some lines - they are not present in annotation data file.  It looks
like the problem is caused by the process that is parsing (lexing) whole
files - because there is no such problem for `diff-annotate dataset ...`.

TODO: fix this bug in annotate.py, add tests.