diffinsights_web: Prepare for introducing Sankey diagram
A Sankey Diagram is a visualisation technique that allows to display
flows: https://www.data-to-viz.com/graph/sankey.html

In the case of PatchScope it would be flow through the directory
structure to the type of changed line at the end.

This commit just reserves space for Sankey diagram in "Contributions:"
widget, though it possibly gets a better name.  There is also
(temporarily) protection added to not crash when it is selected
because it is not implemented yet.

It should be, for the time being, a disabled option, but for some reason
(bug in Panel) it is not disabled, and user is able to select it (!).

https://panel.holoviz.org/reference/widgets/Select.html#core

> - `disabled_options` (list): Optional list of `options` that are disabled,
>    i.e. unusable and un-clickable. If `options` is a dictionary the list
>    items must be dictionary values.