diffinsights_web: Introduce check_is_sankey_possible() method
This method in TimeseriesPlot class (though I am bit unsure if it is a
correct place for it) is intended to be ran automatically via
@param.depend decorator, and manually if needed, to check whether
creating Sankey plot is possible - do we have data (which needs
different file that '*.timeline.*.json' one.

It was planned that this metod woul enable or disable the appopriate
selection option in "Contributions:", but it turns out that accessing
the widget is more complex than at first glance it looks like.