diffinsights_web: Beginnings of mechanism to select plot widget
Now contribution_types_map contains not only the name of the column
to display for plot_commits() (or a special value), but also the
symbolic name for the display widget / pane.  These two pieces of
information are separated by "|" - unlikely to be in column name.

This means that we need to extract plot type from column name in a few
places.  For simplicity, the NO_TEST case does not use the "<plot type>|"
prefix.

To test for correct behavior for unknown, or not supported plots, a
special case of TEST_CASE ("test|TEST") was added.  Few places already
support this.

The value was to be disabled, and is marked as such.  Unfortunately, due
to what looks like bug in Panel, the value is not disabled in HTML (you
can still select it, and it is not grayed out), but there are errors
associated with setting it from UI:

  ERROR: panel.reactive - Callback failed for object named 'Contributions:'
         changing property {'value': 'test|TEST'}
  ValueError: Cannot set the value of Select to 'test|TEST' as it is a disabled option.