diffinsights_web: Add and use .select_plot() in TimeseriesPlot
The new method will be used to select proper plotting function (so that
plot_commits() doesn't get too overloaded), and in the future a proper
plotting pane (plots won't need to use hvPlot / HoloViews, but could
potentailly use Matplotlib / Seaborn, or Plotly).

For TEST_CASE, it returns pn.pane.HTML with message that this plot type
is not supported - as a way of testing getting requiest for a plot that
is not implemented.

Widgets are selected by the way of dict mapping plot type name to
the reactive expression for the plot object.

.
While at it, [temporarily] turn off disabling the TEST_CASE special
type of 'contribution'; it doesn't work anyway.

It will be deleted later in the series; it is here for manual testing.