diffinsights_web: Add seaborn-based heatmap plot for repo
Introduce new type of "Contributions" graph, namely the period-averaged
heatmap of per-commit percentages of types of lines in commits.

This new plot has time on x-axis (sliced into regular periods), and
different types of lines (as categories) on the y-axis.  Lines added
('+') and lines deleted ('-') are analyzed separately: data for lines
added is shown in shades of green, for lines deleted is shown in shades
of red.

The new plot is generated using Seaborn and Matplotlib, and added via
pn.pane.Matplotlib with the format='svg' for responsiveness (see
https://panel.holoviz.org/reference/panes/Matplotlib.html).  This was
done to demonstrate the use of different plotting libraries and
solutions (and for fast prototyping).

This new plot is *not* interactive, because while making plots created
with matplotlib (and seaborn, or any other library using matplotlib as
backend) is possible with `interactive=True`, via interactive ipympl backend
https://matplotlib.org/ipympl/ - in practice it doesn't work so well:
see e.g. https://huggingface.co/spaces/ncusi/demo-panel-plotting-simple

.
TODO: fix what looks like a problem with handling selection of the time
range with the help of "Period:" widgets ("frequency" works fine).

TODO: add support for this heatmap in per-user plots.

TODO: sort types of lines, so that types that usually have the same
prevalence (the same values for percentages in commit) are close
together, via line_type_sorting_key() added for "Line types distribution
[%]" plot).